# api-gateway/swagger.yaml
openapi: 3.0.3
info:
  title: API Gateway - Tugas Besar RPL 2 (Kelompok 3)
  description: |
    Dokumentasi API untuk sistem API Gateway ekosistem UMKM.
    Gateway ini bertindak sebagai proxy dan middleware untuk 6 service lainnya (SmartBank, Marketplace, POS, LogistiKita, SupplierHub, UMKM Insight).
    
    Semua request yang mengarah ke `/api/{service}/*` harus menyertakan **Bearer Token JWT** dan akan dipotong fee 0.5% secara otomatis jika ada parameter finansial.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Gateway Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /auth/token:
    post:
      summary: Generate JWT Token (Dummy)
      description: Membuat JWT token untuk simulasi otentikasi. Token ini wajib digunakan di semua endpoint lain.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: integer
                  example: 1
                username:
                  type: string
                  example: "kelompok3"
                role:
                  type: string
                  example: "admin"
      responses:
        '200':
          description: Token Generated
          content:
            application/json:
              example:
                status: "success"
                message: "Token generated successfully"
                data:
                  token: "eyJhbGciOiJIUzI1NiIsInR..."
...